requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.7.1
MarkupSafe==2.0.1
nltk==3.6.7
numpy==1.21.5
pandas==1.1.5
pandas-datareader==0.10.0
praw==7.5.0
prawcore==2.3.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.1.18
requests==2.27.1
six==1.16.0
textblob==0.17.1
tqdm==4.62.3
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.3
zipp==3.7.0
